Updated MemoryBank.cs
I converted the doubles back to integers and changed the input validation. It closes the activity if the user inputs anything other than lowercased "yes" because it was so difficult figure out how to do a yes/no input validation loop in C# (I know how to do it in Java). It works for now and can be updated later.